move to dio for api calls and fix client.dart request invocations
the http package cannot serialize the body requests that Auth0 requires.
The http package casts all Map body types to <String,String> which does
not support the structure required to be able to set user_metadata.

client.dart was also setup to try and perform a GET, POST, and PATCH all at
the same time by constructing the handler map the way it did. This may have worked
if functions which called the futures were stored in the map, but this is broken
the way it was written.